third_party/WebKit/LayoutTests/http/tests/preload/memcache_eviction_nopreload.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<body>
<script>
var t = async_test('Test that resources get evicted from MemoryCache when preload is not used');
var action;
var loadTimestampWithXHR = t.step_func(function() {
    var xhr = new XMLHttpRequest;
    xhr.withCredentials = true;
    xhr.open("GET", "resources/timestamp.php");
    xhr.send();
    xhr.onload = function() {
        action(xhr.responseText);
    };
});

var writeAndReload = t.step_func(function(text) {
    sessionStorage.timestamp = text;
    document.location.reload(true);
});

var compare = t.step_func(function(text) {
    assert_true(sessionStorage.timestamp != text);
    t.done();
});

if (!sessionStorage.timestamp) {
    action = writeAndReload;
} else {
    action = compare;
}
loadTimestampWithXHR();
</script>
